Implement question 5
Question 5 is asking to implement a function "detectCycle()"
, which can be used to find the entry of cycle inside a linked list.
In addition, the function shouldn't take extra space.
(That means we can only use pointer to detect the cycle.)

The implementation using two pointer, which move through the linked
list at different speed. Because one pointer is faster than other, they
will meet at somewhere in the future if the cycle exist. And the distance
between the meeting point and the head of linked list is same as the distance
between the meeting point and the entry of the cycle. So the only thing
we need to do after these pointers meeting is to move one of them to the
head. Than move these pointers step by step until they meet again.
And finally we'll get the cycle entry.